new: rearchitecture of the updater process Now, BASEDIR is only specified in build_common.sh. Rest of the scripts are automagically modified to source common.sh from BASEDIR as part of the build/script update process.
The updater scripts now execute custom code by using updater-hooks.
There are several points in the lifetime of the updater when such codes
are executed. For now, they are named pre-update-early, pre-update,
pre-update-late, post-update-early and post-update-stage*.

For this hooks mechanism, the updater scripts are now also semi-aware of
the post update stages. They now provide an init.d service
z_updater_post_update, which takes care of executing any
post-update-stage* scripts, of which there can be multiple. After
executing each script, it deletes the script from storage, and performs
a reboot if scheduled by the stage. In the end, when there are no more
stage* scripts available, the service disables and removes itself from
storage.